refactor(net): move all remaining chanel impl specific free functions to `channel.rs` module
